+++ core-ng/src/main/java/core/framework/internal/util/LRUMap.java
@@ -1,4 +1,4 @@
-package core.framework.internal.web.http;
+package core.framework.internal.util;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -6,12 +6,12 @@ import java.util.Map;
 /**
  * @author neo
  */
-class LRUMap<K, V> extends LinkedHashMap<K, V> {
+public class LRUMap<K, V> extends LinkedHashMap<K, V> { // not thread safe
     private static final long serialVersionUID = -191933814768129858L;
 
     private final int maxSize;
 
-    LRUMap(int maxSize) {
+    public LRUMap(int maxSize) {
         super(maxSize, 0.75F, true);
         this.maxSize = maxSize;
     }