+++ ext/monitor/src/main/java/Main.java
@@ -1,0 +1,10 @@
+import app.MonitorApp;
+
+/**
+ * @author ericchung
+ */
+public class Main {
+    public static void main(String[] args) {
+        new MonitorApp().start();
+    }
+}