+++ ext/monitor/src/main/java/app/MonitorApp.java
@@ -1,0 +1,50 @@
+package app;
+
+import app.monitor.ActionAlertConfig;
+import app.monitor.action.ActionAlertService;
+import app.monitor.kafka.ActionLogMessageHandler;
+import app.monitor.kafka.EventMessageHandler;
+import app.monitor.slack.SlackClient;
+import app.monitor.slack.SlackMessageAPIRequest;
+import app.monitor.slack.SlackMessageAPIResponse;
+import core.framework.http.HTTPClient;
+import core.framework.http.HTTPClientBuilder;
+import core.framework.json.Bean;
+import core.framework.log.message.ActionLogMessage;
+import core.framework.log.message.EventMessage;
+import core.framework.log.message.LogTopics;
+import core.framework.module.App;
+import core.framework.module.SystemModule;
+
+/**
+ * @author ericchung
+ */
+public class MonitorApp extends App {
+    public static final String MONITOR_APP = "monitor";
+
+    @Override
+    protected void initialize() {
+        load(new SystemModule("sys.properties"));
+        loadProperties("app.properties");
+
+        configureSlackClient(); // currently only support slack notification
+
+        configureActionAlert();
+    }
+
+    private void configureActionAlert() {
+        Bean.register(ActionAlertConfig.class);
+        ActionAlertConfig config = Bean.fromJSON(ActionAlertConfig.class, requiredProperty("app.actionAlertConfig"));
+        bind(new ActionAlertService(config));
+        kafka().subscribe(LogTopics.TOPIC_ACTION_LOG, ActionLogMessage.class, bind(ActionLogMessageHandler.class));
+        kafka().subscribe(LogTopics.TOPIC_EVENT, EventMessage.class, bind(EventMessageHandler.class));
+    }
+
+    private void configureSlackClient() {
+        bind(HTTPClient.class, new HTTPClientBuilder().build());
+
+        Bean.register(SlackMessageAPIRequest.class);
+        Bean.register(SlackMessageAPIResponse.class);
+        bind(new SlackClient(requiredProperty("app.slackToken")));
+    }
+}