+++ ext/monitor/src/main/java/app/monitor/slack/SlackMessageAPIResponse.java
@@ -1,0 +1,13 @@
+package app.monitor.slack;
+
+import core.framework.api.json.Property;
+import core.framework.api.validate.NotNull;
+
+/**
+ * @author ericchung
+ */
+public class SlackMessageAPIResponse {
+    @NotNull
+    @Property(name = "ok")
+    public Boolean ok;
+}