+++ ext/monitor/src/test/java/app/ConfigTest.java
@@ -1,0 +1,15 @@
+package app;
+
+import org.junit.jupiter.api.Test;
+
+import static core.framework.test.Assertions.assertConfDirectory;
+
+/**
+ * @author ericchung
+ */
+class ConfigTest extends IntegrationTest {
+    @Test
+    void conf() {
+        assertConfDirectory().overridesDefaultResources();
+    }
+}